+++ config/environment.rb
@@ -12,6 +12,7 @@ end
 # some timeslot variables, put them here until we need a better place
 STRANGER = 'stranger'
 FRIEND = 'friend'
+FOAF = 'foaf'
 ME = 'me'
 PAGINATION_DEFAULT_LIMIT = ENV['PAGINATION_DEFAULT_LIMIT'].try(:to_i) || 40
 PAGINATION_MAX_LIMIT = ENV['PAGINATION_MAX_LIMIT'].try(:to_i) || 100