+++ spec/services/presentable_slots_spec.rb
@@ -30,6 +30,16 @@ RSpec.describe PresentableSlots, type: :service do
           described_class.call(relationship: FRIEND, user: user,
                                current_user: current_user)
         ).to eq [user.std_slots_public, user.std_slots_friends,
+                 user.std_slots_foaf, user.re_slots,
+                 current_user.shared_group_slots(user)]
+
+      end
+
+      it "for a friend-of-a-friend (foaf)" do
+        expect(
+          described_class.call(relationship: FOAF, user: user,
+                               current_user: current_user)
+        ).to eq [user.std_slots_public, user.std_slots_foaf,
                  user.re_slots, current_user.shared_group_slots(user)]
 
       end